use renderComponent and reorganize interfaces
