+++ redisinsight/ui/src/pages/browser/components/add-key/AddKeyVectorSet/AddKeyVectorSet.types.ts
@@ -1,7 +1,16 @@
 import { Maybe } from 'uiSrc/utils'
+import { PopulateMode } from './constants'
 
 export interface Props {
   keyName: string
   keyTTL: Maybe<number>
   onCancel: (isCancelled?: boolean) => void
+}
+
+export interface PopulateOption {
+  value: PopulateMode
+  label: string
+  description?: string
+  disabled?: boolean
+  id: string
 }
